pyproject.toml
[project]
name = "planform-backend"
version = "0.1.0"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "sqlalchemy>=2",
  "asyncpg",
  "redis>=5",
  "openai>=1",
  "playwright==1.*",
  "python-dotenv",
  "pydantic-settings"
]
